role

+++ backend/routes/api.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\middleware\CheckRole;
 
 /*
 |--------------------------------------------------------------------------
@@ -19,13 +20,13 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 });
 
 
-Route::get("/clients", "App\Http\Controllers\ClientController@getAllClients");
-Route::get("/clients/{id}", "App\Http\Controllers\ClientController@getClientById");
-Route::get("/softdeletedclients", "App\Http\Controllers\ClientController@getSoftDeletedClients");
-Route::post("/clients", "App\Http\Controllers\ClientController@createClient");
-Route::delete("/clients/{id}", "App\Http\Controllers\ClientController@deleteClient");
-Route::post("/clients/{id}/restore", "App\Http\Controllers\ClientController@restoreClient");
-Route::put("/clients/{id}", "App\Http\Controllers\ClientController@updateClient");
-Route::get("/clients/{id}/image", "App\Http\Controllers\ClientController@serveImage");
+Route::get("/clients", "App\Http\Controllers\ClientController@getAllClients")->middleware(CheckRole::class . ":doctor, administator");
+Route::get("/clients/{id}", "App\Http\Controllers\ClientController@getClientById")->middleware(CheckRole::class . ":doctor, administator, client");
+Route::get("/softdeletedclients", "App\Http\Controllers\ClientController@getSoftDeletedClients")->middleware(CheckRole::class . ":doctor, administator");
+Route::post("/clients", "App\Http\Controllers\ClientController@createClient")->middleware(CheckRole::class . ":doctor, administator");
+Route::delete("/clients/{id}", "App\Http\Controllers\ClientController@deleteClient")->middleware(CheckRole::class . ":doctor, administator");
+Route::post("/clients/{id}/restore", "App\Http\Controllers\ClientController@restoreClient")->middleware(CheckRole::class . ":doctor, administator");
+Route::put("/clients/{id}", "App\Http\Controllers\ClientController@updateClient")->middleware(CheckRole::class . ":doctor, administator");
+Route::get("/clients/{id}/image", "App\Http\Controllers\ClientController@serveImage")->middleware(CheckRole::class . ":doctor, administator, client");
 Route::post("/login", "App\Http\Controllers\UserController@login");
-Route::post("/register", "App\Http\Controllers\UserController@register");
+Route::post("/register", "App\Http\Controllers\UserController@register")->middleware(CheckRole::class . ":doctor, administator");
